+++ src/Problems/f_leading_ones_ruggedness3.hpp
@@ -8,7 +8,7 @@
 #ifndef _F_LEADING_ONES_RUGGEDNESSTHREE_H
 #define _F_LEADING_ONES_RUGGEDNESSTHREE_H
 
-#include "../IOHprofiler_problem.hpp"
+#include "../Template/IOHprofiler_problem.hpp"
 #include "common_used_functions/wmodels.hpp"
 
 class LeadingOnes_Ruggedness3 : public IOHprofiler_problem<int> {
@@ -57,7 +57,7 @@ public:
       }
     }
     result = IOHprofiler_get_evaluate_double_info()[(int)(result+0.5)];
-    y.push_back((double)result);
+    y.push_back(result);
     return y;
   };
